--- a/pkg/store/opts.go
+++ b/pkg/store/opts.go
@@ -27,7 +27,14 @@ func init() {
 func findSetMatches(pattern string) []string {
 	escaped := false
 	sets := []*strings.Builder{{}}
-	for i := 0; i < len(pattern); i++ {
+	init := 0
+	end := len(pattern)
+	// If the regex is wrapped in a group we can remove the first and last parentheses
+	if pattern[init] == '(' && pattern[end-1] == ')' {
+		init++
+		end--
+	}
+	for i := init; i < end; i++ {
 		if escaped {
 			switch {
 			case isRegexMetaCharacter(pattern[i]):