--- a/pkg/store/opts_test.go
+++ b/pkg/store/opts_test.go
@@ -24,6 +24,15 @@ func TestFindSetMatches(t *testing.T) {
 				"baz",
 			},
 		},
+		// Simple sets with group wrapper.
+		{
+			pattern: "(foo|bar|baz)",
+			exp: []string{
+				"foo",
+				"bar",
+				"baz",
+			},
+		},
 		// Simple sets containing escaped characters.
 		{
 			pattern: "fo\\.o|bar\\?|\\^baz",
@@ -36,6 +45,10 @@ func TestFindSetMatches(t *testing.T) {
 		// Simple sets containing special characters without escaping.
 		{
 			pattern: "fo.o|bar?|^baz",
+			exp:     nil,
+		},
+		{
+			pattern: "(fool|bar)|(baz)",
 			exp:     nil,
 		},
 		{